refactor: refactored to ddd layout & removed empty tests
