--- bgg/client_test.go
@@ -1,1 +1,0 @@
-package bgg
