+++ bgg/thing.go
@@ -2,27 +2,6 @@ package bgg
 
 import "encoding/xml"
 
-type Search struct {
-	XMLName    xml.Name `xml:"items"`
-	Text       string   `xml:",chardata"`
-	Total      string   `xml:"total,attr"`
-	Termsofuse string   `xml:"termsofuse,attr"`
-	Item       []struct {
-		Text string `xml:",chardata"`
-		Type string `xml:"type,attr"`
-		ID   string `xml:"id,attr"`
-		Name struct {
-			Text  string `xml:",chardata"`
-			Type  string `xml:"type,attr"`
-			Value string `xml:"value,attr"`
-		} `xml:"name"`
-		Yearpublished struct {
-			Text  string `xml:",chardata"`
-			Value string `xml:"value,attr"`
-		} `xml:"yearpublished"`
-	} `xml:"item"`
-}
-
 type Thing struct {
 	XMLName    xml.Name `xml:"items" json:"items,omitempty"`
 	Text       string   `xml:",chardata" json:"text,omitempty"`
